requirements.txt
mcp[cli]>=1.0.0
pymysql>=1.0.2
pandas>=1.5.0
numpy>=1.20.0
scikit-learn>=1.0.0
python-dotenv>=0.19.0
pydantic>=1.10.0
requests>=2.28.0 
openai>=1.66.3
uv>=0.6.8